pyproject.toml
[project]
name = "ask-your-files"
version = "0.1.0"
description = ""
authors = [
    {name = "tanatip_timtong",email = "ttimt@sertiscorp.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "typing-extensions (>=4.14.1,<5.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "dynaconf (>=3.2.11,<4.0.0)",
    "langchain-text-splitters (>=0.3.9,<0.4.0)",
    "langchain-chroma (>=0.2.5,<0.3.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "marker-pdf (>=1.8.5,<2.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.31,<0.4.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "lark (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
